+++ crates/miden-protocol/src/account/account_id/seed.rs
@@ -2,7 +2,7 @@ use alloc::vec::Vec;
 
 use crate::account::account_id::AccountIdVersion;
 use crate::account::account_id::v1::{compute_digest, validate_prefix};
-use crate::account::{AccountIdV1, AccountStorageMode};
+use crate::account::{AccountIdV1, AccountType};
 use crate::errors::AccountError;
 use crate::{Felt, Word};
 
@@ -14,14 +14,14 @@ use crate::{Felt, Word};
 /// implementation was removed in commit dab6159318832fc537bb35abf251870a9129ac8c in PR 1061.
 pub(super) fn compute_account_seed(
     init_seed: [u8; 32],
-    storage_mode: AccountStorageMode,
+    account_type: AccountType,
     version: AccountIdVersion,
     code_commitment: Word,
     storage_commitment: Word,
 ) -> Result<Word, AccountError> {
     compute_account_seed_single(
         init_seed,
-        storage_mode,
+        account_type,
         version,
         code_commitment,
         storage_commitment,
@@ -30,7 +30,7 @@ pub(super) fn compute_account_seed(
 
 fn compute_account_seed_single(
     init_seed: [u8; 32],
-    storage_mode: AccountStorageMode,
+    account_type: AccountType,
     version: AccountIdVersion,
     code_commitment: Word,
     storage_commitment: Word,
@@ -47,14 +47,14 @@ fn compute_account_seed_single(
 
     // loop until we have a seed that satisfies the specified account parameters.
     loop {
-        // Check if the seed satisfies the specified type, storage mode and version. Additionally,
-        // the most significant bit of the suffix must be zero to ensure felt validity.
+        // Check if the seed satisfies the specified account type and version. Additionally, the
+        // most significant bit of the suffix must be zero to ensure felt validity.
         let suffix = current_digest[AccountIdV1::SEED_DIGEST_SUFFIX_ELEMENT_IDX];
         let prefix = current_digest[AccountIdV1::SEED_DIGEST_PREFIX_ELEMENT_IDX];
         let is_suffix_msb_zero = suffix.as_canonical_u64() >> 63 == 0;
 
-        if let Ok((computed_storage_mode, computed_version)) = validate_prefix(prefix)
-            && computed_storage_mode == storage_mode
+        if let Ok((computed_account_type, computed_version)) = validate_prefix(prefix)
+            && computed_account_type == account_type
             && computed_version == version
             && is_suffix_msb_zero
         {